examples/requirements.txt
# Requirements for agent automation examples
PyGithub>=2.1.1
python-dotenv>=1.0.0
